backend/requirements.txt
fastapi
uvicorn
pydantic
google-genai
tavily-python
python-dotenv
wbgapi
